SE-299 Changed directory of parity checker and cleanup
