+++ sdk/src/test/tutorial/java/com/finbourne/lusid/tutorials/ibor/Reconciliation.java
@@ -6,12 +6,11 @@ import com.finbourne.lusid.api.InstrumentsApi;
 import com.finbourne.lusid.api.ReconciliationsApi;
 import com.finbourne.lusid.api.TransactionPortfoliosApi;
 import com.finbourne.lusid.model.*;
-import com.finbourne.lusid.tutorials.ibor.parity.LusidFeature;
+import parity.LusidFeature;
 import com.finbourne.lusid.utilities.*;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
-import java.io.Console;
 import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
 import java.util.ArrayList;
