+++ sdk/src/test/tutorial/java/com/finbourne/lusid/tutorials/ibor/Valuation.java
@@ -6,7 +6,7 @@ import com.finbourne.lusid.api.InstrumentsApi;
 import com.finbourne.lusid.api.QuotesApi;
 import com.finbourne.lusid.api.TransactionPortfoliosApi;
 import com.finbourne.lusid.model.*;
-import com.finbourne.lusid.tutorials.ibor.parity.LusidFeature;
+import parity.LusidFeature;
 import com.finbourne.lusid.utilities.*;
 import org.junit.BeforeClass;
 import org.junit.Test;
