+++ sdk/src/test/tutorial/java/parity/ParityFileWriter.java
@@ -1,4 +1,4 @@
-package com.finbourne.lusid.tutorials.ibor.parity;
+package parity;
 
 import java.io.*;
